Reorganize scripts into disassemblers, tags, and analysis
Group auxiliary scripts by purpose: classic FLOSS tool integrations under
disassemblers/, QuantumStrand tag DB maintenance under tags/ (including
build_oss_db.py), and batch runs under analysis/. Add category READMEs,
update tests and import paths, and point build_oss_db default output at
floss/qs/db/data/oss.
