+++ floss/language/rust/rust_version_database.py
@@ -25,8 +25,8 @@
 #
 # To regenerate or update this file, you can follow these steps:
 # 1. Navigate to the script directory.
-# 2. Execute the script 'extract_rust_hashes.py'.
-#    Example command: python extract_rust_hashes.py
+# 2. Execute scripts/extract_rust_hashes.py from the repository root.
+#    Example command: python scripts/extract_rust_hashes.py
 #############################################################################################
