+++ scripts/README.md
@@ -1,33 +1,38 @@
 # FLOSS Scripts
-FLOSS supports converting its output into scripts for various tools. Please see the render scripts in this directory.
-  
-Additionally, there is another [plugin for IDA](idaplugin.py) to allow FLOSS to automatically
-extract obfuscated strings and apply them to the currently loaded module in IDA. `idaplugin.py` is a IDAPython script you can directly run within IDA Pro (File - Script File... [ALT + F7]).
 
-# Installation
-These scripts can be downloaded from the FLOSS [GitHub](https://github.com/mandiant/flare-floss) repository
-alongside the source, which is required for the scripts to run.
-To install FLOSS as source, see the documentation [here](../doc/installation.md).
+Auxiliary scripts live under `scripts/`, grouped by purpose:
 
+| Directory | Purpose |
+|-----------|---------|
+| [`disassemblers/`](disassemblers/) | Convert classic FLOSS JSON output into import scripts for Binary Ninja, Ghidra, IDA Pro, Radare2, and x64dbg; includes the IDA plugin |
+| [`tags/`](tags/) | Build and maintain QuantumStrand tag databases (global prevalence, OSS libraries, VT feeds) |
+| [`analysis/`](analysis/) | Batch analysis helpers |
 
-# Usage
-## Convert FLOSS output for use by other tools
+## disassemblers/
 
-- Run FLOSS on the desired executable with the `-j` or `--json` argument to emit a JSON result
-and redirect it to a JSON file.  
-    `$ floss -j suspicious.exe > floss_results.json`
+Turn FLOSS JSON (`floss -j sample.exe > results.json`) into tool-specific artifacts.
 
-For Binary Ninja, IDA Pro, Ghidra or Radare2:
-- Run the script for your tool of choice by passing the result json file as an argument and
-redirect the output to a Python (.py) file.  
+1. Run a render script, redirecting stdout to a file.
+2. Import or run the generated artifact in the target tool.
 
-Ghidra Example:  
-    `$ python render-ghidra-import-script.py floss_results.json > apply_floss.py`
+Example (Ghidra):
 
-- Run the Python script `apply_floss.py` using the desired tool.
+```console
+$ python scripts/disassemblers/render-ghidra-import-script.py results.json > apply_floss.py
+```
 
-For x64dbg:
-- Instead of a Python file, redirect the output to a .json file.  
-    `$ python render-x64dbg-database.py floss-results.json > database.json`
+See [`disassemblers/`](disassemblers/) for per-tool scripts and the [IDA plugin](disassemblers/idaplugin.py) (`File → Script file…` in IDA Pro).
 
-- Open the JSON file `database.json` in x64dbg.
+Install FLOSS from source first; see [installation](../doc/installation.md).
+
+## tags/
+
+Scripts that extract strings, build tag databases, and query them. See [`tags/README.md`](tags/README.md) for the full pipeline.
+
+## analysis/
+
+- [`bulk_analyze.py`](analysis/bulk_analyze.py) — run `floss` or `floss quantum` over every binary in a directory and write JSON results.
+
+## Other
+
+- [`extract_rust_hashes.py`](extract_rust_hashes.py) — regenerate `floss/language/rust/rust_version_database.py` from Rust release metadata (not part of the FLOSS or QuantumStrand runtimes).
