+++ scripts/analysis/bulk_analyze.py
