+++ scripts/disassemblers/render-r2-import-script.py
@@ -23,7 +23,7 @@ that marks up the current workspace.
 Usage:
 
   $ floss suspicious.exe -j > floss-results.json
-  $ python render-r2-import-script.py floss-results.json > apply_floss.py
+  $ python scripts/disassemblers/render-r2-import-script.py floss-results.json > apply_floss.py
   # now run `apply_floss.py` in radare2
 """
